osx/requirements.txt
# FastAPI and server dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0

# Stable Diffusion and ML dependencies - Mac compatible
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=0.27.0
safetensors>=0.4.0

# Image processing
Pillow>=10.2.0

# HTTP client for generator script
requests>=2.31.0

# PyTorch is installed separately in setup_mac.sh
# to ensure proper Metal support 